More FB.BX_BOXING_IMMEDIATELY_UNBOXED_TO_PERFORM_COERCION in RealSingle

+++ src/unikl/disco/numbers/implementations/RealSingle.java
@@ -187,7 +187,7 @@ public class RealSingle implements Num {
 	
 	@Override
 	public double doubleValue() {
-	    return new Double( value ).doubleValue();
+	    return new Double( value );
 	}
 
 	@Override
